move text from experimental to components dir
