+++ packages/components/text/package.json
@@ -26,9 +26,11 @@
     "directory": "packages/components/text"
   },
   "dependencies": {
-    "@uifabricshared/foundation-compose": "^1.12.0",
     "@fluentui-react-native/adapters": ">=0.9.0 <1.0.0",
-    "@fluentui-react-native/tokens": ">=0.16.0 <1.0.0"
+    "@fluentui-react-native/framework": "0.8.0",
+    "@fluentui-react-native/theme-tokens": ">=0.19.0 <1.0.0",
+    "@fluentui-react-native/tokens": ">=0.16.0 <1.0.0",
+    "tslib": "^2.3.1"
   },
   "devDependencies": {
     "@fluentui-react-native/eslint-config-rules": "^0.1.1",
