+++ packages/components/text/src/Variants.win32.ts
