+++ packages/components/text/src/deprecated/Text.settings.ts
