+++ packages/components/text/src/index.ts
@@ -1,3 +1,22 @@
-export { textName } from './Text.types';
-export type { ITextProps, ITextType } from './Text.types';
-export { Text } from './Text';
+export { Text as TextV1 } from './Text';
+export type { textName as textNameV1, TextProps, TextTokens } from './Text.types';
+export {
+  Caption1,
+  Body1,
+  Body1Strong,
+  Body2,
+  Body2Strong,
+  Subtitle1,
+  Subtitle1Strong,
+  Subtitle2,
+  Subtitle2Strong,
+  Title1,
+  Title1Strong,
+  LargeTitle,
+  Display,
+} from './Variants';
+
+/* deprecated */
+export { textName } from './deprecated/Text.types';
+export type { ITextProps, ITextType } from './deprecated/Text.types';
+export { Text } from './deprecated/Text';
