--- packages/experimental/Text/jest.config.js
@@ -1,2 +1,0 @@
-const { configureReactNativeJest } = require('@fluentui-react-native/scripts');
-module.exports = configureReactNativeJest('android');
