+++ packages/experimental/Text/package.json
@@ -28,6 +28,7 @@
   "dependencies": {
     "@fluentui-react-native/adapters": ">=0.9.0 <1.0.0",
     "@fluentui-react-native/framework": "0.8.0",
+    "@fluentui-react-native/text": ">=0.12.0 <1.0.0",
     "@fluentui-react-native/theme-tokens": ">=0.19.0 <1.0.0",
     "tslib": "^2.3.1"
   },
